Implemented FastCDC
This allows for variable and meaningful chunk sizes with good anomaly resistance
and good similarity when only small changes are added. In addition this should also have
good rolling similarity, so if a file gets shifted the algorithm should be better than power of 2
chunking.

TODO: Implement on the controller side, it is all currently skipped.
TODO: see what "special case" is when skipping files or folders. Fix it?